backend/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-rq==2.8.1
django-axes==7.0.1
psycopg2-binary==2.9.11
Pillow==11.2.1
numpy==2.2.5
pandas==2.2.3
scipy==1.15.3
scikit-learn==1.6.1
imbalanced-learn==0.13.0
joblib==1.5.0
requests==2.32.3
redis==7.3.0
rq==1.16.2
langdetect==1.0.9
presidio-analyzer==2.2.361
presidio-anonymizer==2.2.361
tensorflow==2.20.0
torch==2.10.0
torchvision==0.25.0
medmnist==3.0.2
scikit-image==0.25.2

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
